screen: Restore terminal when using ctrl+z
This drove me nuts. Whenever using ctrl+z in e.g. vim, I could not see
the previous contents of the terminal. Pressing ctr+z to drop out of vim
to test whatever I'm working on before hopping back on is super
convinient. And if any error messages pop up, I'd like to be able to
jump into and out of vim without losing those messages. For some reason
the terminal was cleared on some systems, and restored on others. Turned
out to be GNU screen, fixable with altscreen=on

Credits:https://unix.stackexchange.com/questions/528040/how-to-control-screen-behavior-when-suspending-vim-with-ctrl-z#comment977025_528040