xinput: Disable fancy input methods
A fresh installation of Debian buster included and automatically enabled
Ibus, Intelligent Input Bus. It's supposed to be smart, but it really
gets in the way a lot. Tilde activated some prompt, which is really
inconvenient when trying to refer to your homedir in a terminal.
Checking the settings and searching the web revealed no solutions, other
than to switch to a different input method. I've disabled it all.